Move existing status to separate module
