--- src/lib.rs
@@ -1,10 +1,0 @@
-pub mod central_dispatch;
-pub mod gtfs_fetcher;
-pub mod messages;
-pub mod objects;
-pub mod session_actor;
-pub mod siri_stuff;
-pub mod templates;
-pub mod status;
-
-use objects::*;
