+++ src/main.rs
@@ -6,15 +6,15 @@ use actix::prelude::*;
 use actix_web::{get, middleware, web, App, HttpRequest, HttpResponse, HttpServer, Responder};
 use actix_web_actors::ws;
 use color_eyre::{eyre::format_err, Result};
-use idfm_proxy::central_dispatch::CentralDispatch;
-use idfm_proxy::gtfs_fetcher::GtfsFetcher;
-use idfm_proxy::messages::SiriUpdate;
-use idfm_proxy::objects::{LineReference, StopReference};
-use idfm_proxy::session_actor::{SessionActor, Watching};
-use idfm_proxy::siri_stuff::SiriFetcher;
-use idfm_proxy::status::status;
-use idfm_proxy::templates;
+use objects::{LineReference, StopReference};
 use tracing_subscriber::{filter::targets::Targets, layer::SubscriberExt, util::SubscriberInitExt};
+
+mod actors;
+mod messages;
+mod objects;
+mod status;
+mod templates;
+use actors::*;
 
 #[get("/ws")]
 async fn websocket(
@@ -153,7 +153,7 @@ async fn main() -> color_eyre::Result<()> {
                     .flat_map(|frame| frame.estimated_vehicle_journey)
             })
             .collect();
-        dispatch_addr.do_send(SiriUpdate { vjs });
+        dispatch_addr.do_send(messages::SiriUpdate { vjs });
         tracing::info!("Re-using old data");
     }
 
@@ -167,7 +167,7 @@ async fn main() -> color_eyre::Result<()> {
     }
     .start();
 
-    let _gtfs_fetch = GtfsFetcher {
+    let _gtfs_fetch = actors::GtfsFetcher {
         dispatch: dispatch_addr.clone(),
     }
     .start();
@@ -179,7 +179,7 @@ async fn main() -> color_eyre::Result<()> {
             .service(actix_files::Files::new("/static", "./static"))
             .service(index)
             .service(line)
-            .service(status)
+            .service(status::status)
             .service(websocket)
             .service(line_websocket)
     })
